feat: persisting columns
