+++ src/modules/exporter/__init__.py
@@ -1,6 +1,9 @@
 """Defines the module's public interface."""
 
+from .columns import load_columns, set_columns
 from .exporter import export_to_csv
 from .journal_parser import journal_responses_to_data_frame
+from .results_table_header import ResultsTableHeader
 
-__all__ = ["export_to_csv", "journal_responses_to_data_frame"]
+__all__ = ["export_to_csv", "journal_responses_to_data_frame", "load_columns",
+           "set_columns", "ResultsTableHeader"]
